add login

+++ dwitter/controllers/dweets.js
@@ -1,4 +1,4 @@
-import * as dweetRepository from '../data/data.js';
+import * as dweetRepository from '../data/dweet.js';
 
 export const getDweets = async (req, res, next) => {
   const username = req.query.username;
@@ -17,8 +17,8 @@ export const getDweetById = async (req, res, next) => {
 };
 
 export const postDweet = async (req, res, next) => {
-  const { text, username, name } = req.body;
-  const newDweet = await dweetRepository.create(text, username, name);
+  const { text } = req.body;
+  const newDweet = await dweetRepository.create(text, req.userId);
   res.status(201).json(newDweet);
 };
 
@@ -27,15 +27,22 @@ export const updateDweet = async (req, res, next) => {
   const { text } = req.body;
   const dweet = await dweetRepository.update(id, text);
   if (dweet) {
-    dweet.text = text;
-    res.status(200).json(dweet);
+    if (req.userId !== id) {
+      return res.sendStatus(403);
+    } else {
+      dweet.text = text;
+      return res.status(200).json(dweet);
+    }
   } else {
-    res.status(404).json({ message: `Dweet id(${id}) not found!` });
+    return res.status(404).json({ message: `Dweet id(${id}) not found!` });
   }
 };
 
 export const deleteDweet = async (req, res, next) => {
   const { id } = req.params;
-  await dweetRepository.remove(id);
-  res.sendStatus(204);
+  if (req.userId !== id) {
+    return res.sendStatus(403);
+  }
+  dweetRepository.remove(id);
+  return res.sendStatus(204);
 };
